% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLeagueTable.R
\name{CreateLeagueTable}
\alias{CreateLeagueTable}
\title{Create a League table for the selected matches}
\usage{
CreateLeagueTable(Matches, date = c(""))
}
\arguments{
\item{Matches}{dataframe a datafame of matches obtained from StasBombR functions get.matches or FreeMatches}

\item{date}{vector a string in the format YYYY-MM-DD. it can be a selection between 2 dates. see example}
}
\value{
a leage table dataframe
}
\description{
Create a League table for the selected matches
}
\examples{
library(StatsBombR)
Competition <- FreeCompetitions() |>
  filter(competition_name == "Serie A", season_name == "2015/2016")
  Matches <- FreeMatches(Competitions = Competition)
## match date as date
Matches <- Matches |>
  mutate(match_date = as.Date(match_date),
         match_id = as.factor(match_id))
## whole season league table
CreateLeagueTable(Matches)

## partial season table
CreateLeagueTable(Matches, date = c("2026-01-01"))

## season table in between two dates
CreateLeagueTable(Matches, date = c("2026-01-01", 2016-05-26))
}
