% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PossessionMatch.R
\name{PossessionMatch}
\alias{PossessionMatch}
\title{Create a League table for the selected matches}
\usage{
PossessionMatch(MatchEvents)
}
\arguments{
\item{MatchEvents}{A datafame of Matches Events obtained from StasBombR functions free_allevents. multiple matches are allowed}
}
\value{
a dataframe with the amount of passees attempted and the percentage possiseesion for each teeam
}
\description{
Create a League table for the selected matches
}
